# docs/openapi/private-backend.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Me 2 Backend API",
    "description": "Backend Server API endpoints for Me 2 - Node.js/Express",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/api/memories/{memoryId}": {
      "get": {
        "tags": ["Public", "Private"],
        "summary": "GET /api/memories/{memoryId}",
        "description": "Get a specific memory by ID",
        "operationId": "GetMemoryById",
        "parameters": [
          {
            "description": "The ID of the memory to retrieve",
            "in": "path",
            "name": "memoryId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Memory object",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "404": {
            "description": "Memory not found",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": {} }
            }
          }
        },
        "security": []
      },
      "patch": {
        "tags": ["Private"],
        "summary": "PATCH /api/memories/{memoryId}",
        "description": "/api/memories/{memoryId}\nUpdate an existing memory",
        "operationId": "UpdateMemory",
        "parameters": [
          {
            "description": "The ID of the memory to update",
            "in": "path",
            "name": "memoryId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateMemoryRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated memory object",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "404": {
            "description": "Memory not found",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": {} }
            }
          }
        },
        "security": []
      },
      "delete": {
        "tags": ["Private"],
        "summary": "DELETE /api/memories/{memoryId}",
        "description": "/api/memories/{memoryId}?mode=\"\"\nDelete a memory (soft or hard delete)",
        "operationId": "DeleteMemory",
        "parameters": [
          {
            "description": "The ID of the memory to delete",
            "in": "path",
            "name": "memoryId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "description": "Delete mode: 'soft' (default, marks as deleted) or 'hard' (completely removes)",
            "in": "query",
            "name": "mode",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["soft", "hard"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Deletion confirmation",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "404": {
            "description": "Memory not found",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": {} }
            }
          }
        },
        "security": []
      }
    },
    "/api/memories": {
      "post": {
        "tags": ["Private"],
        "summary": "POST /api/memories",
        "description": "/api/memories\nCreate a new memory",
        "operationId": "CreateMemory",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateMemoryRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Newly created memory object",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "400": {
            "description": "Missing required fields",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": {} }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/activities": {
      "get": {
        "tags": ["Private"],
        "summary": "GET /api/activities",
        "description": "/api/activities\nGet all activities with their questions",
        "operationId": "GetActivities",
        "parameters": [],
        "responses": {
          "200": {
            "description": "All activities and their associated questions",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": {} }
            }
          }
        },
        "security": []
      }
    },
    "/api/activities/status/{userId}": {
      "get": {
        "tags": ["Private"],
        "summary": "GET /api/activities/status/{userId}",
        "description": "/api/activities/status/:userId\nGet the status of all activities for a specific user\nStatus includes: not_started, in_progress, completed",
        "operationId": "GetActivityStatus",
        "parameters": [
          {
            "description": "- The user ID to get activity statuses for",
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Activity statuses for the user",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": {} }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/memories/user/{userId}": {
      "get": {
        "tags": ["Private"],
        "summary": "GET /api/memories/user/{userId}",
        "description": "Get all memories for a specific user",
        "operationId": "GetMemoriesByUser",
        "parameters": [
          {
            "description": "The user ID to fetch memories for",
            "in": "path",
            "name": "userId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "description": "Whether to include soft-deleted memories (default: false)",
            "in": "query",
            "name": "includeDeleted",
            "required": false,
            "schema": { "type": "boolean" }
          }
        ],
        "responses": {
          "200": {
            "description": "Array of memory objects",
            "content": {
              "application/json": { "schema": {} }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": { "schema": {} }
            }
          }
        },
        "security": []
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateMemoryRequest": {
        "properties": {
          "userId": { "type": "string" },
          "type": { "type": "string", "enum": ["user", "ai"] },
          "content": { "type": "string" },
          "source": { "type": "string" },
          "categories": { "items": { "type": "string" }, "type": "array" },
          "sectionIds": { "items": { "type": "string" }, "type": "array" }
        },
        "required": ["userId", "type", "content"],
        "type": "object",
        "additionalProperties": true
      },
      "UpdateMemoryRequest": {
        "properties": {
          "content": { "type": "string" },
          "source": { "type": "string" },
          "categories": { "items": { "type": "string" }, "type": "array" }
        },
        "type": "object",
        "additionalProperties": true
      },
...
    },
    "securitySchemes": {}
  },
  "servers": [{ "url": "/" }]
}